feat(build): add root package and feature flags for solver and store
- Introduce a new root package `wildside-engine` with optional features:
  - `solver-vrp` (default)
  - `solver-ortools` (optional stub solver crate added)
  - `store-sqlite` (optional SQLite POI store)
- Gate `SqlitePoiStore` and solver usage behind feature flags in `wildside-core` and `wildside-cli`
- Forward features from member crates via `dep:` dependencies for a single source of truth
- Add CLI errors and tests for missing features to handle no-default-features builds gracefully
- Remove monolithic `wildside-core::store` module; split into submodules under feature flag
- Introduce tests and behavioural scenarios exercising feature selection and fallback
- Update documentation and roadmap to reflect feature flag strategy and implementation

This enables flexible builds and prepares for future expansion of solver and store backends.
