+++ wildside-cli/tests/features/feature_flags.feature
@@ -1,0 +1,6 @@
+Feature: Feature flags
+
+  Scenario: Ingest requires the store-sqlite feature
+    Given valid ingest inputs exist
+    When I run the ingest command
+    Then the command fails because store-sqlite is disabled
