+++ wildside-solver-ortools/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "wildside-solver-ortools"
+version = "0.1.0"
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+rust-version.workspace = true
+publish = false
+
+[dependencies]
+wildside-core = { workspace = true }
+
+[lints]
+workspace = true
